Make the internal Slice methods crate-private
There's no need to expose raw buffer access methods of Slice
in the public API.
